+++ R/analysis_functions.R
@@ -26,7 +26,8 @@ analyze_factor <- function(experiment, analysis_factor) {
     group2 = c(),
     group3 = c(),
     group4 = c(),
-    group5 = c()
+    group5 = c(),
+    group6 = c(),
   )
 
   names(expt_groups) <- unique(names(group_organized))
@@ -103,3 +104,48 @@ analyze_factor <- function(experiment, analysis_factor) {
       subtitle = paste0("Factor being compared: ", analysis_factor)
     )
 }
+
+
+#' @title Generate ANOVA tables for Detected Calls
+#'
+#' @description Aggregates raw data tables for all data added to experiment object,
+#' and calculates and displays ANOVA statistics.
+#'
+#' @param experiment The experiment object
+#' @param analysis_factor A string representing the factor to analyze between groups
+#'
+#' @return A bar graph comparing the analysis_factor between groups
+#'
+#' @examples \dontrun{squeakrANOVA(experiment = experiment, analysis_factor = "Delta_Freq")}
+#'
+#' @import dplyr
+#' @import rlist
+#' @importFrom rstatix anova_test
+#' @export
+squeakrANOVA <- function(experiment, analysis_factor) {
+  group_organized <- list.group(experiment$experimental_data, group)
+
+  combined_df <- data.frame(matrix(ncol = ncol(experiment$experimental_data[1]$call_data$raw), nrow = 0))
+
+  for (selected_group in 1:length(names(group_organized))) {
+    for (dataset in 1:length(group_organized[[selected_group]])) {
+      set_to_add <- group_organized[[selected_group]][dataset]$call_data$raw %>%
+        mutate(group = names(group_organized)[selected_group])
+      combined_df <- rbind(combined_df, set_to_add)
+    }
+  }
+
+  combined_df <- combined_df %>%
+    mutate(index = 1:nrow(combined_df))
+  colnames(combined_df) <- trimws(gsub(r"{\s*\([^\)]+\)}","", as.character(names(combined_df))))
+  colnames(combined_df) <- gsub(" ", "_", names(combined_df), fixed = TRUE)
+  res.aov <- anova_test(data = combined_df, dv = analysis_factor, wid = index, between = group)
+  return(get_anova_table(res.aov))
+}
+
+
+
+
+
+
+